Added a dialogue system & Installed TMPro
The dialogue system reads text with ajustable speeds, allows for easy implementation of multiple dialogue lines, and since it's a prefab it can easily be placed in different scenes and have different triggers enable it!

It requires TextMeshPro to work as such this has also been installed in this commit.